% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_ontoProc.R
\name{demoPhenoPlot}
\alias{demoPhenoPlot}
\title{demoPhenoPlot is a function to do a demo of taking a few phenotypes defined herein and using ontoProc to visualize}
\usage{
demoPhenoPlot(showDemo = T)
}
\arguments{
\item{showDemo}{set to T to}
}
\description{
demoPhenoPlot is a function to do a demo of taking a few phenotypes defined herein and using ontoProc to visualize
}
