% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_ontoProc.R
\name{fix.onto.tag}
\alias{fix.onto.tag}
\title{fix.onto.tag is a function that fixes CL_ to CL:}
\usage{
fix.onto.tag(onto.tag)
}
\arguments{
\item{onto.tag}{give it a tag}
}
\description{
fix.onto.tag is a function that fixes CL_ to CL:
}
